==> 1. NNS on ratings.ipynb <==
import numpy as np
import pandas as pd
from scipy.spatial.distance import pdist, squareform

ratings = pd.read_csv('ratings.csv')

def NNSrecommend(inputuser = 1, ratings=ratings):
    
    """Suggest movies based on NNS algorihm.
    
    Input:
    inputuser       -- "userId" of the selected user
    
    Output:
    recommendations -- "movieId"-s of recommended movies, numpy.ndarray format
    
    Error codes:
    -1              -- "inputuser" not found in database
    -2              -- There are no users, that have rated the same movies
    """
    
    if inputuser not in ratings['userId'].unique():
        return -1

    # Reduction of movies
    seenmovies = ratings[ratings['userId'] == inputuser]['movieId']
    smallratings = ratings[ratings['movieId'].isin(seenmovies)]

    # Reduction of users
    users = smallratings['userId'].value_counts() == len(seenmovies)
    users = users.index[users].tolist()
    smallratings = smallratings[smallratings['userId'].isin(users)]
        
    # Make pivot table and calculate distances
    table = pd.pivot_table(smallratings, values='rating', index='userId', columns='movieId')

    distances = pdist(table, metric='euclidean')
    distances = squareform(distances)
    
    # Select the distances that we need
    distances = pd.DataFrame(index = table.index, columns = table.index, data = distances)
    distances = distances[distances.index == inputuser].drop(inputuser, axis = 1)
    
    if distances.empty:
        return -2
    
    nearestuser = distances.idxmin(axis = 1).values[0]

    recommendations = ratings[(ratings['userId'] == nearestuser) & (ratings['movieId'].isin(seenmovies) == False)
                              & (ratings['rating'] >= 4)]['movieId'].values
    
    return recommendations

NNSrecommend(1)

NNSrecommend(436344)

# Coverage

suggested_movies = []
users = np.random.choice(ratings['userId'].unique(), size=100)
for i in users:
    suggested_movies.append(NNSrecommend(i))

list = []
for i in range(len(suggested_movies)):
    if type(suggested_movies[i]) == np.ndarray:
        for j in range(len(suggested_movies[i])):
            list.append(suggested_movies[i][j])

len(np.unique(list))

ratings.movieId.nunique()

# Personalization

from sklearn.metrics.pairwise import cosine_similarity

suggested_movies = []
users = np.random.choice(ratings['userId'].unique(), size=100)
for i in users:
    tmp = NNSrecommend(i)
    if type(tmp) != int:
        suggested_movies.append(tmp)

list = []
for i in range(len(suggested_movies)):
    for j in range(len(suggested_movies[i])):
        list.append(suggested_movies[i][j])

list = np.unique(list)

data = np.zeros(shape=(len(suggested_movies), len(list)))
seen_movies = pd.DataFrame(data=data, columns=list)

for i in range(len(suggested_movies)):
    for j in suggested_movies[i]:
        if j in seen_movies.columns:
            seen_movies.loc[i,j] = 1

pd.DataFrame(cosine_similarity(seen_movies))

# Test values
x = seen_movies.iloc[0]
y = seen_movies.iloc[1]

sum(x * y) / ( sum(x**2)**(1/2) * sum(y**2)**(1/2) )

matrix = pd.DataFrame(cosine_similarity(seen_movies))

matrix = 1-matrix

personalization = []
for i in range(matrix.shape[0]):
    for j in range(matrix.shape[1]):
        if i < j:
            personalization.append(matrix.loc[i,j])

np.mean(personalization)
