==> 11. Global movie recommender with K-Prototypes clustering.ipynb <==
import numpy as np
import pandas as pd
from sklearn.preprocessing import MultiLabelBinarizer
from sklearn.preprocessing import MinMaxScaler

from kmodes.kprototypes import KPrototypes

import time
import pickle

# Drop ratings of previously dropped movies
# ratings = pd.read_csv('ratings.csv')
# ratings = ratings[ratings['movieId'].isin(pd.read_csv('3b movies_GloVe.csv')['id'])]
# ratings.to_csv('2 ratings.csv', index=False)

# Delete duplicates and write to file
# unprepared_movies = pd.read_csv('3b movies_GloVe.csv')
# unprepared_movies.drop_duplicates(inplace=True)
# duplicates = (unprepared_movies['id'].value_counts() > 1).head(13).index
# unprepared_movies = unprepared_movies[unprepared_movies['id'].isin(duplicates) == False]
# unprepared_movies.to_csv('4 movies.csv', index=False)

movies = pd.read_csv('4 movies.csv')
unprepared_movies = pd.read_csv('4 movies.csv')
country_codes = pd.read_csv('country_codes.csv')
ratings = pd.read_csv('2 ratings.csv')

def get_numbers(str):
    """Return the integer numbers from a string in an array format."""
    arr = str.replace(',','').replace('{','').replace('}','').replace('[','').replace(']','').split()
    arr = [int(s) for s in arr if s.isdigit()]
    return arr

def get_country_codes(str):
    """Return the codes of the countries from a string."""
    arr = str.replace(',','').replace('{','').replace('}','').replace('[','').replace(']','').replace("'", "").split()
    arr = [s for s in arr if country_codes['Code'].str.contains(s).any()]
    return arr

def prepare_dataframe(movies):
    
    # Conversion
    movies['genres'] = movies['genres'].apply(lambda str: get_numbers(str))
    movies['production_companies'] = movies['production_companies'].apply(lambda str: get_numbers(str))
    movies['production_countries'] = movies['production_countries'].apply(lambda str: get_country_codes(str)) # This takes time
    
    # Get dummies
    mlb = MultiLabelBinarizer()
    genres = pd.DataFrame(mlb.fit_transform(movies['genres']), columns=mlb.classes)

    mlb = MultiLabelBinarizer()
    production_companies = pd.DataFrame(mlb.fit_transform(movies['production_companies']), columns=mlb.classes)

    mlb = MultiLabelBinarizer()
    production_countries = pd.DataFrame(mlb.fit_transform(movies['production_countries']), columns=mlb.classes)

    original_language = pd.get_dummies(movies['original_language'])
    
    # Rename the columns to unique names
    for i in range(genres.shape[1]):
        genres.rename(mapper={i : str(i) + '. genre'}, axis=1, inplace=True)
    
    for i in range(production_countries.shape[1]):
        production_countries.rename(mapper={i : str(i) + '. country'}, axis=1, inplace=True)
    
    movieids = movies['id']
    movies = movies.drop(['id', 'title'], axis=1)
    
    # production_companies, original languages, production countries are not used, because of huge running time
    numeric = movies.drop(['adult', 'genres', 'original_language', 'production_companies', 'production_countries'], axis=1)
    categorical = movies[['adult']].join(genres)
    
    # Scale the numeric values to (0, 1) interval
    scaler = MinMaxScaler()
    numeric = scaler.fit_transform(numeric)
    
    movievectors = pd.DataFrame(numeric).join(categorical)
    movievectors.index = movieids
    
    return movievectors

def convert_title_to_index(x):
    row = unprepared_movies2[unprepared_movies2['title'] == x]
    return np.array(row['id'])[0]

movies = prepare_dataframe(movies)

Train model and save it:

# Sampling used to reduce running time
# movies_sample = movies.sample(n=4000, random_state=98)

# start = time.time()
# kproto = KPrototypes(n_clusters=45, max_iter=20, n_jobs=4)
# kproto.fit(movies_sample, categorical=[i for i in range(57,movies_sample.shape[1])])
# end = time.time()
# elapsed = end-start

# elapsed

# pickle.dump(kproto, open("kproto_4000.p", 'wb'))

Calcute the clusters of all movies to save time in recommending:

# all_movies_clusters = kproto.predict(movies, categorical=[i for i in range(57,movies.shape[1])])
# pickle.dump(all_movies_clusters, open("clusters_kproto_4000.p", 'wb'))

# Recommend movies:

kproto = pickle.load(open("kproto_4000.p", "rb"))
all_movies_clusters = pickle.load(open("clusters_kproto_4000.p", "rb"))

def recommend_movies(uid, unprepared_movies, movies, ratings, cluster_obj, clusters):
    """
    Input:
    uid - ID of the user
    
    Output:
    potential_suggestions - The most popular 20 suggested movies
    
    Error codes:
    -1 - This userID does not exist
    -2 - There are no movies in the current form of movies database that was rated by the user
    """
    
    rated_movies = ratings[ratings['userId'] == uid].drop('userId', axis=1)

    if rated_movies.shape[0] == 0:
        return -1 # Invalid userId

    best_match = find_best_matching_cluster(uid, movies, rated_movies, cluster_obj)
    
    if best_match == -2:
        return -2 # There is no rated movie in movies database

    # Select the movies in the best matching cluster
    potential_suggestions = unprepared_movies[unprepared_movies['id'].isin(movies[clusters == best_match].index)]

    # Delete previously watched (rated) movies
    watched_index = potential_suggestions[potential_suggestions['id'].isin(rated_movies['movieId'])].index
    potential_suggestions = potential_suggestions.drop(watched_index)

    # First suggest the more popular movies
    potential_suggestions = potential_suggestions.sort_values('popularity', ascending=False)

    # Grab the first twenty
    potential_suggestions = potential_suggestions['title'].head(20)
    
    return potential_suggestions

def find_best_matching_cluster(uid, movies, rated_movies, cluster_obj):
    # The clusters of movies that have been rated at least once
    clusters = np.array([])
    
    # The indexes of movies that have not been found in movies database
    # Which indicates error in data set
    indexlist = np.array([])
    
    for index, row in rated_movies.iterrows():
        try:
            clusters = np.append(clusters, cluster_obj.predict(pd.DataFrame(movies.loc[int(row['movieId'])]).transpose(), categorical=[i for i in range(57,movies.shape[1])]))
        except:
            indexlist = np.append(indexlist, index)
    
    # Handle inconsistency in DB
    rated_movies.drop(indexlist, inplace=True)

    rated_movies['clusters'] = clusters

    # Find the highest rated cluster
    if rated_movies.shape[0] > 0:
        best_match = rated_movies[['rating', 'clusters']].groupby('clusters').mean().idxmax()[0]
    else:
        return -2 # There is no rated movie in movies database
    
    return best_match

recommend_movies(5000, unprepared_movies, movies, ratings, kproto, all_movies_clusters)

recommend_movies(235623464, unprepared_movies, movies, ratings, kproto, all_movies_clusters)

# Coverage

suggested_movies = []
users = np.random.choice(ratings['userId'].unique(), size=100)
for i in users:
    suggested_movies.append(recommend_movies(i, unprepared_movies, movies, ratings, kproto, all_movies_clusters))

list = []
for i in range(len(suggested_movies)):
    if type(suggested_movies[i]) == pd.core.series.Series:
        for j in suggested_movies[i].index:
            list.append(j)

len(np.unique(list))

ratings.movieId.nunique()

# Personalization

from sklearn.metrics.pairwise import cosine_similarity

suggested_movies = []
users = np.random.choice(ratings['userId'].unique(), size=100)
for i in users:
    tmp = recommend_movies(i, unprepared_movies, movies, ratings, kproto, all_movies_clusters)
    if type(tmp) != int:
        suggested_movies.append(tmp)

list = []
for i in range(len(suggested_movies)):
    for j in suggested_movies[i]:
        list.append(j)
        
list = np.unique(list)

data = np.zeros(shape=(len(suggested_movies), len(list)))
seen_movies = pd.DataFrame(data=data, columns=list)

for i in range(len(suggested_movies)):
    for j in suggested_movies[i]:
        if j in seen_movies.columns:
            seen_movies.loc[i,j] = 1

matrix = pd.DataFrame(cosine_similarity(seen_movies))

matrix = 1-matrix

matrix

personalization = []
for i in range(matrix.shape[0]):
    for j in range(matrix.shape[1]):
        if i < j:
            personalization.append(matrix.loc[i,j])

np.mean(personalization)

# Intra-list similarity

movies.drop(movies.iloc[:, 0:58], axis=1, inplace=True)

# For suggestion 0
matrix = pd.DataFrame(cosine_similarity(movies[movies.index.isin(suggested_movies[0].index)]))

intra_list = []
for i in range(matrix.shape[0]):
    for j in range(matrix.shape[1]):
        if i < j:
            intra_list.append(matrix.loc[i,j])

np.mean(intra_list)

# For all 100 suggestions
intra_list_100 = []
for i in suggested_movies:
    matrix = pd.DataFrame(cosine_similarity(movies[movies.index.isin(i.index)]))
    
    intra_list = []
    for i in range(matrix.shape[0]):
        for j in range(matrix.shape[1]):
            if i < j:
                intra_list.append(matrix.loc[i,j])
    
    intra_list_100.append(np.mean(intra_list))

np.mean(intra_list_100)
